shared/pyproject.toml
[project]
name = "app-shared"
version = "0.1.0"
description = "Shared models and config for ETL Studio backend and worker"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app_shared"]
